refactor: created data/Number, data/String, data/Object
