+++ src/data/Object.ts
@@ -1,0 +1,27 @@
+/**
+ * @since 1.0.0
+ */
+
+import { pipe } from "@fp-ts/core/Function"
+import * as I from "@fp-ts/schema/internal/common"
+import type { AnnotationOptions, Schema } from "@fp-ts/schema/Schema"
+
+/**
+ * @since 1.0.0
+ */
+export const instanceOf = <A extends abstract new(...args: any) => any>(
+  constructor: A,
+  annotationOptions?: AnnotationOptions<object>
+) =>
+  (self: Schema<object>): Schema<InstanceType<A>> =>
+    pipe(
+      self,
+      I.filter(
+        (a): a is InstanceType<A> => a instanceof constructor,
+        {
+          description: `an instance of ${constructor.name}`,
+          custom: { type: "instanceOf", instanceOf: constructor },
+          ...annotationOptions
+        }
+      )
+    )
