+++ test/data/Option.ts
@@ -2,7 +2,7 @@ import { pipe } from "@fp-ts/core/Function"
 import * as O from "@fp-ts/core/Option"
 import * as S from "@fp-ts/schema"
 import * as _ from "@fp-ts/schema/data/Option"
-import { parseNumber } from "@fp-ts/schema/data/parser"
+import { parseNumber } from "@fp-ts/schema/data/String"
 import * as P from "@fp-ts/schema/Parser"
 import * as Pretty from "@fp-ts/schema/Pretty"
 import * as Util from "@fp-ts/schema/test/util"
