+++ test/data/ReadonlySet.ts
@@ -1,6 +1,6 @@
 import { pipe } from "@fp-ts/core/Function"
-import { parseNumber } from "@fp-ts/schema/data/parser"
 import * as _ from "@fp-ts/schema/data/ReadonlySet"
+import { parseNumber } from "@fp-ts/schema/data/String"
 import * as P from "@fp-ts/schema/Parser"
 import * as Pretty from "@fp-ts/schema/Pretty"
 import * as S from "@fp-ts/schema/Schema"
