+++ test/data/filter/Int.ts
@@ -1,5 +1,5 @@
 import { pipe } from "@fp-ts/core/Function"
-import * as _ from "@fp-ts/schema/data/filter"
+import * as _ from "@fp-ts/schema/data/Number"
 import * as P from "@fp-ts/schema/Parser"
 import * as Pretty from "@fp-ts/schema/Pretty"
 import * as S from "@fp-ts/schema/Schema"
