+++ test/data/filter/MinLength.ts
@@ -1,4 +1,4 @@
-import * as _ from "@fp-ts/schema/data/filter"
+import * as _ from "@fp-ts/schema/data/String"
 import * as P from "@fp-ts/schema/Parser"
 import * as Pretty from "@fp-ts/schema/Pretty"
 import * as S from "@fp-ts/schema/Schema"
