+++ test/data/filter/StartsWith.ts
@@ -1,5 +1,4 @@
 import { pipe } from "@fp-ts/core/Function"
-import * as _ from "@fp-ts/schema/data/filter"
 import * as P from "@fp-ts/schema/Parser"
 import * as S from "@fp-ts/schema/Schema"
